+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - `Ferrum::Browser#close` closes browser gracefully issuing a CDP command, doesn't clean up ruby resources.
 - `Ferrum::Node#remove` removes node from DOM tree.
 - `Ferrum::Node#exists?` check whether the node in ruby world still exists in the DOM tree.
+- `Ferrum::Cookies#store` stores all cookies of current page in a file.
+- `Ferrum::Cookies#load` Loads all cookies from the file and sets them for current page.
 
 ### Changed